fix(executor-agent-rs): use busybox base and fix musl ioctl type
- Dockerfile: use busybox:latest instead of scratch — init container
  needs `cp` to copy the binary into the user container. Remove
  .cargo/config.toml at build time since Alpine already targets musl.
- pty_util: cast TIOCSCTTY to c_int (musl ioctl signature differs from
  glibc).

Verified: 11/11 SDK integration tests pass on kind cluster.

+++ rust/executor-agent/src/pty_util.rs
@@ -44,7 +44,8 @@ pub fn setup_slave_terminal(slave_fd: i32) -> io::Result<()> {
         .map_err(|e| io::Error::other(format!("setsid: {}", e)))?;
 
     // Set controlling terminal
-    let ret = unsafe { libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0i32) };
+    #[allow(clippy::unnecessary_cast)]
+    let ret = unsafe { libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_int, 0i32) };
     if ret < 0 {
         return Err(io::Error::last_os_error());
     }
